feat(day4): 消除代码重复 + 路由拆分 + AppState 同步桥接
- src/api/routes/: 5 个路由模块 (health/qa/insight/evaluate/crawl)
- src/api/main.py: FastAPI 应用组装
- api.py: 向后兼容重定向到 src.api.main
- AppState: 新增 init_sync() / rebuild_sync() 供 Streamlit 使用
- app.py: 删除独立的 build_runtime/reload_after_fetch → 委托给 AppState
- src/runtime.py: 标记 deprecated（功能已被 AppState 取代）
